requirements.txt
fastapi==0.103.1
py-eureka-client==0.11.7
uvicorn==0.23.2
faiss-cpu==1.7.4
FlagEmbedding
python-multipart